refactor: created models for user
